+++ app/src/main/java/com/walkingforrochester/walkingforrochester/android/ui/composable/contact/ContactUsScreen.kt
@@ -1,10 +1,10 @@
-package com.walkingforrochester.walkingforrochester.android.ui.composable.faq
+package com.walkingforrochester.walkingforrochester.android.ui.composable.contact
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.walkingforrochester.walkingforrochester.android.ui.composable.common.WFRWebView
 
 @Composable
-fun FAQScreen(modifier: Modifier = Modifier) {
+fun ContactUsScreen(modifier: Modifier = Modifier) {
     WFRWebView(modifier = modifier, url = "https://walkingforrochester.org/faq/")
 }
